updated indent and wrap guide colors
I noticed before that the colors used in the indent guide and wrap guide seemed really dark, so I lightened them in the last release. On closer inspection, they seemed dark because the indent guide in ST3 is actually a subtle dotted line, rather than solid. So I tweaked the guide colors in this theme to match them visually.

Let me know how it looks!